+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-## XX.XX.XX 
+## XX.XX.XX
+* ! Minor breaking change ! All active views will now automatically stop when consent for "views" is revoked.
+
 * Added new functions to ease the presenting the feedback widgets. Functions present the first matching feedback widget from the list.
   * presentNPS(Context)
   * presentNPS(Context, String)